refactor: 공통 시간 엔티티와 JPA 감사 설정 추가
