+++ src/main/java/org/sopt/global/config/JpaAuditConfig.java
@@ -1,0 +1,9 @@
+package org.sopt.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditConfig {
+}
